+++ src/protocols/writing.rs
@@ -13,8 +13,8 @@ use tracing::*;
 
 use std::{any::Any, collections::HashMap, io, net::SocketAddr};
 
-/// Can be used to specify and enable writing, i.e. sending outbound messages.
-/// If handshaking is enabled too, it goes into force only after the handshake has been concluded.
+/// Can be used to specify and enable writing, i.e. sending outbound messages. If the `Handshake`
+/// protocol is enabled too, it goes into force only after the handshake has been concluded.
 #[async_trait]
 pub trait Writing: Pea2Pea
 where